Workaround onLayout triggered constantly on some devices, fixes #50
This was observed reliably on a Nexus 5X with React Native 0.31+
onLayout would be triggered constantly with slightly different width and
height for the thumb size when its position changes.
As a result it triggered a re-render constantly.
This will need to be investigated on the React Native side as it's not
an expected behavior.